GEODE-6143: remove PowerMock from HeapMemoryMonitorTest
